some things on functios

+++ FunctionsLib.h
@@ -4,6 +4,6 @@ namespace Elevator{
   class Functions{
       public:
           // Returns a + b
-          static double Add(int a, int b);
+          static int Add(int a, int b);
   };
 }
